transform_load/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
google-cloud-core